requirements.txt
Django==1.8.4
argparse==1.2.1
django-webpack-loader==0.2.4
djangorestframework==3.3.2
ipython==5.3.0
pathlib2==2.2.1
pickleshare==0.7.4
six==1.10.0
traitlets==4.3.2
wheel==0.26.0
wsgiref==0.1.2
